+++ management/CHANGELOG.md
@@ -1,4 +1,4 @@
-# v0.8.0 (Unreleased)
+# v0.8.0 (2022-09-04)
 
 * **Bug fix** - Fix `Gateway` response dereferencing for non-`LDAP` types [#40](https://github.com/patrickcping/pingone-go-sdk-v2/pull/40)
 * **Bug fix** - Fix `ExportCSR` headers, to determine the response type of the CSR [#41](https://github.com/patrickcping/pingone-go-sdk-v2/pull/41)